Question on Part 3
It looks like I'm almost done.  I'll just need to test the various
filters with a constant and then plug in the variable from your main
page.

However, my guess is that the resetMap function isn't actually clearing
the map when I push the Plot Data button.  I know bc of the console.log
is going through all the steps, and it is adding markers without
clearing the old ones.

I've tried a few returns to grab the "plotted" variable unsuccessfully.

My main question is how do I get that resetMap function to actually grab
the value of all the markers I created?